Removed onConfirm callback for usePreventReload
